Cambios en el script de compilacion

+++ Practica3/frontwrapper/copyToTomcat.sh
@@ -10,6 +10,8 @@ static_dir="$wrapper_dir/src/main/resources/static"
 # Script
 npm run build
 
-rm -rf $static_dir/*
-
-cp -r $dist_dir/* $static_dir
+# If compilation returns exit code = 0 then copy dist to wrapper's static directory
+if [ $? -eq 0 ]; then
+    rm -rf $static_dir/*
+    cp -r $dist_dir/* $static_dir   
+fi
